requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
clickhouse-driver==0.2.5
clickhouse-sqlalchemy==0.2.3
idna==3.4
kafka-python==2.0.2
numpy==1.24.2
pandas==1.5.3
pyarrow==11.0.0
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
